requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
